+++ cloudwanderer/storage_connectors/dynamodb.py
@@ -7,7 +7,7 @@ import os
 import pathlib
 from functools import reduce
 from random import randrange
-from typing import Callable, Iterable, Iterator, List
+from typing import Any, Callable, Dict, Iterable, Iterator, List
 
 import boto3
 from boto3.dynamodb.conditions import Attr, ConditionBase, Key
@@ -87,20 +87,33 @@ def _dynamodb_items_to_resources(items: Iterable[dict], loader: Callable) -> Ite
         loader (Callable): The method which can be used to fulfil the :meth:`CloudWandererResource.load`
 
     """
-    for item_id, group in itertools.groupby(items, lambda x: x["_id"]):
+    for _, group in itertools.groupby(items, lambda x: x["_id"]):
         grouped_items = list(group)
         attributes = [
-            SecondaryAttribute(name=attribute["_attr"], **attribute)
+            SecondaryAttribute(name=attribute["_attr"], **_strip_dynamodb_attrs(attribute))
             for attribute in grouped_items
             if attribute["_attr"] != "BaseResource"
         ]
-        base_resource = next(iter(resource for resource in grouped_items if resource["_attr"] == "BaseResource"))
+        base_resource = next(resource for resource in grouped_items if resource["_attr"] == "BaseResource")
+        parent_urn = URN.from_string(base_resource["_parent_urn"]) if base_resource.get("_parent_urn") else None
+        subresource_urns = [URN.from_string(urn) for urn in base_resource.get("_subresource_urns", [])]
         yield CloudWandererResource(
             urn=_urn_from_primary_key(base_resource["_id"]),
-            resource_data=base_resource,
+            parent_urn=parent_urn,
+            subresource_urns=subresource_urns,
+            resource_data=_strip_dynamodb_attrs(base_resource),
             secondary_attributes=attributes,
             loader=loader,
         )
+
+
+def _strip_dynamodb_attrs(raw_dict: Dict[str, Any]) -> Dict[str, Any]:
+    """Remove any underscore prefixed keys as these are attributes we use to identify the DynamoDB record.
+
+    Arguments:
+        raw_dict: The raw dictionary of the DynamoDB record that needs cleaning.
+    """
+    return {k: v for k, v in raw_dict.items() if not k.startswith("_")}
 
 
 class DynamoDbConnector(BaseStorageConnector):
@@ -169,9 +182,12 @@ class DynamoDbConnector(BaseStorageConnector):
     def write_resource(self, resource: CloudWandererResource) -> None:
         logger.debug(f"Writing: {resource.urn} to {self.table_name}")
         item = {
-            **self._generate_index_values_for_write(resource.urn),
+            **self._generate_urn_index_values(resource.urn),
             **standardise_data_types(resource.cloudwanderer_metadata.resource_data or {}),
+            **{"_subresource_urns": [str(urn) for urn in resource.subresource_urns]},
         }
+        if resource.parent_urn:
+            item["_parent_urn"] = str(resource.parent_urn)
         self.dynamodb_table.put_item(Item=item)
         for secondary_attribute in resource.cloudwanderer_metadata.secondary_attributes:
             self._write_secondary_attribute(
@@ -191,12 +207,12 @@ class DynamoDbConnector(BaseStorageConnector):
         """
         logger.debug(f"Writing: {attribute_type} of {urn} to {self.table_name}")
         item = {
-            **self._generate_index_values_for_write(urn, attribute_type),
+            **self._generate_urn_index_values(urn, attribute_type),
             **standardise_data_types(secondary_attribute or {}),
         }
         self.dynamodb_table.put_item(Item=item)
 
-    def _generate_index_values_for_write(self, urn: URN, attr: str = "BaseResource") -> dict:
+    def _generate_urn_index_values(self, urn: URN, attr: str = "BaseResource") -> dict:
         values = {
             "_id": _primary_key_from_urn(urn),
             "_attr": attr,
@@ -225,8 +241,15 @@ class DynamoDbConnector(BaseStorageConnector):
         result = self.dynamodb_table.query(KeyConditionExpression=Key("_id").eq(_primary_key_from_urn(urn)))
         return next(_dynamodb_items_to_resources(result["Items"], loader=self.read_resource), None)
 
-    def read_resources(self, **kwargs) -> Iterator["CloudWandererResource"]:
-        query_generator = DynamoDbQueryGenerator(**kwargs)
+    def read_resources(
+        self,
+        account_id: str = None,
+        region: str = None,
+        service: str = None,
+        resource_type: str = None,
+        urn: URN = None,
+    ) -> Iterator["CloudWandererResource"]:
+        query_generator = DynamoDbQueryGenerator(account_id, region, service, resource_type, urn)
         for condition_expression in query_generator.condition_expressions:
             query_args = {"Select": "ALL_PROJECTED_ATTRIBUTES", "KeyConditionExpression": condition_expression}
             if query_generator.index is not None:
@@ -249,6 +272,9 @@ class DynamoDbConnector(BaseStorageConnector):
         resource_records = self.dynamodb_table.query(KeyConditionExpression=Key("_id").eq(_primary_key_from_urn(urn)))[
             "Items"
         ]
+        resource_records += self.dynamodb_table.query(
+            IndexName="parent_urn", KeyConditionExpression=Key("_parent_urn").eq(str(urn))
+        )["Items"]
         with self.dynamodb_table.batch_writer() as batch:
             for record in resource_records:
                 logger.debug("Deleting %s", record["_id"])