+++ tests/integration/helpers.py
@@ -9,6 +9,7 @@ import moto
 from jmespath.lexer import LexerError
 from moto import ec2
 
+from cloudwanderer import URN
 from cloudwanderer.cloud_wanderer_resource import CloudWandererResource
 
 DEFAULT_SESSION = boto3.Session(aws_access_key_id="11111111", aws_secret_access_key="111111", aws_session_token="1111")
@@ -120,7 +121,10 @@ class GenericAssertionHelpers:
             KeyError: Occurs when the key does not exist
         """
         try:
-            return str(getattr(resource, key))
+            value = getattr(resource, key)
+            if isinstance(value, URN):
+                return str(value)
+            return value
         except AttributeError:
             try:
                 return resource.get_secondary_attribute(jmes_path=key)
@@ -150,6 +154,8 @@ class SetupMocking:
             "cloudwanderer.aws_interface.CloudWandererAWSInterface.limit_resources", new=self.limit_resources
         )
         self.service_mocks = {}
+        self.regions_mock = PropertyMock(return_value=[])
+        self.regions_patcher = patch("moto.ec2.models.RegionsAndZonesBackend.regions", new=self.regions_mock)
         clear_aws_credentials()
 
     def start_general_mock(
@@ -157,10 +163,7 @@ class SetupMocking:
     ):
         restrict_regions = ["eu-west-2", "us-east-1"] if restrict_regions is None else restrict_regions
         if restrict_regions:
-            ec2.models.RegionsAndZonesBackend.regions = [
-                ec2.models.Region(region_name, "ec2.{region_name}.amazonaws.com", "opt-in-not-required")
-                for region_name in restrict_regions
-            ]
+            self.start_restrict_regions(regions=restrict_regions)
         if restrict_services:
             self.start_restrict_services(services=["ec2", "s3", "iam"])
         if limit_resources is not False:
@@ -168,7 +171,12 @@ class SetupMocking:
         self.start_moto_services()
 
     def stop_general_mock(self):
-        stop_methods = [self.stop_moto_services, self.stop_restrict_services, self.stop_limit_collections_list]
+        stop_methods = [
+            self.stop_moto_services,
+            self.stop_restrict_services,
+            self.stop_limit_collections_list,
+            self.stop_restrict_regions,
+        ]
         for method in stop_methods:
             try:
                 method()
@@ -208,6 +216,16 @@ class SetupMocking:
 
     def stop_limit_collections_list(self):
         self.collections_patcher.stop()
+
+    def start_restrict_regions(self, regions):
+        self.regions_mock.return_value = [
+            ec2.models.Region(region_name, "ec2.{region_name}.amazonaws.com", "opt-in-not-required")
+            for region_name in regions
+        ]
+        self.regions_patcher.start()
+
+    def stop_restrict_regions(self):
+        self.regions_patcher.stop()
 
 
 DEFAULT_MOCKER = None